refactor: HttpRequest 리팩토링
